requirements.txt
torch
openai-whisper
pyannote.audio
numpy
streamlit